requirements.txt
line-bot-sdk
bs4
flask
pymongo
pandas
scipy
azure-cognitiveservices-vision-computervision